Move story config
